Added experimental WebSocket motor control system
I spent the better part of the last two days trying to make it as robust and easy to work with as possible, but in the end I probably never needed it. Oh well, it's a pretty cool thing to implement anyway.
I separated some parts of the auth to make it easier to authenticate without needing a full HTTP request, and tweaked the Python motor controller to work like one expects.